Fixed bugs in intalling and loading songs

+++ src/controller/installer/Installer.java
@@ -103,7 +103,7 @@ public class Installer implements InstallerInterface {
      */
     private void installMusic() {
         for (final Music m : Music.values()) {
-            if (!Files.exists(Paths.get(Folder.MUSICFOLDER.getAbsolutePath() + m.getAbsolutePath()), LinkOption.NOFOLLOW_LINKS)) {
+            if (m != Music.SONG && !Files.exists(Paths.get(Folder.MUSICFOLDER.getAbsolutePath() + m.getAbsolutePath()), LinkOption.NOFOLLOW_LINKS)) {
                 try(InputStream musicStream = this.getClass().getResourceAsStream(m.getResourcePath());
                         FileOutputStream fos = new FileOutputStream(Folder.MUSICFOLDER.getAbsolutePath() + m.getAbsolutePath())) {
                     final byte[] buf = new byte[SIZE];
